@@ -1,72 +1,86 @@
 {
-  "name": "react-spreadsheet-component",
+  "name": "react-sheets",
   "description": "A Spreadsheet Component for ReactJS",
-  "version": "0.5.1",
+  "homepage": "https://github.com/patrick-jones/react-sheets",
+  "version": "0.1.0",
   "license": "MIT",
-  "repository": "felixrieseberg/React-Spreadsheet-Component",
-  "maintainers": [
-    {
-      "name": "Felix Rieseberg",
-      "email": "felix@felixrieseberg.com",
-      "web": "http://www.felixrieseberg.com"
-    }
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/patrick-jones/react-sheets.git"
+  },
+  "author": "Patrick Jones <pjnola@gmail.com> (https://github.com/patrick-jones)",
+  "contributors": [
+    "Patrick Jones <pjnola@gmail.com> (https://github.com/patrick-jones)",
+    "Felix Rieseberg <felix@felixrieseberg.com> (http://www.felixrieseberg.com)"
   ],
   "dependencies": {
-    "jquery": "^2.1",
-    "mousetrap": "^1.5.3"
+    "classnames": "^2.2.5",
+    "core-js": "^2.4.1"
   },
   "peerDependencies": {
-    "react": "^0.14.3"
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2"
   },
   "devDependencies": {
-    "body-parser": "^1.12.4",
-    "browserify": "^12.0.1",
-    "browserify-shim": "^3.8.11",
-    "del": "^2.0.2",
-    "express": "^4.12.3",
-    "gulp": "^3.8.11",
-    "gulp-connect": "^4.0.0",
-    "gulp-header": "^1.2.2",
-    "gulp-jshint": "^1.9.4",
-    "gulp-plumber": "^1.0.0",
-    "gulp-react": "^3.0.1",
-    "gulp-rename": "^1.2.0",
-    "gulp-streamify": "1.0.2",
-    "gulp-uglify": "^1.1.0",
-    "gulp-util": "^3.0.4",
-    "jest-cli": "^0.8.2",
-    "jshint-stylish": "^2.0.1",
-    "multer": "^1.1.0",
-    "react": "^0.14.0",
-    "react-addons-test-utils": "^0.14.3",
-    "react-dom": "^0.14.0",
-    "react-tools": "^0.13.3",
-    "reactify": "^1.1.1",
-    "tape": "^4.2.2",
-    "vinyl-source-stream": "^1.1.0"
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.18.2",
+    "babel-eslint": "^7.1.0",
+    "babel-loader": "^6.2.7",
+    "babel-plugin-react-transform": "^2.0.2",
+    "babel-polyfill": "^6.16.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-2": "^6.18.0",
+    "cpx": "^1.5.0",
+    "cross-env": "^3.1.3",
+    "css-loader": "^0.25.0",
+    "eslint": "^3.10.0",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-react": "^6.6.0",
+    "expect": "^1.20.2",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "karma": "^1.3.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-cli": "^1.0.1",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^1.8.0",
+    "phantomjs-polyfill": "0.0.2",
+    "phantomjs-prebuilt": "^2.1.13",
+    "postcss": "^5.2.5",
+    "postcss-cssnext": "^2.8.0",
+    "postcss-loader": "^1.1.1",
+    "precss": "^1.4.0",
+    "react": "^15.3.2",
+    "react-addons-test-utils": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "react-hot-loader": "^3.0.0-beta.6",
+    "style-loader": "^0.13.1",
+    "stylelint": "^7.5.0",
+    "stylelint-config-standard": "^14.0.0",
+    "webpack": "^1.13.3"
   },
   "scripts": {
-    "debug": "gulp --debug",
-    "dist": "gulp bundle-js --production --release && gulp bundle-js --development --release",
-    "watch": "gulp",
-    "test": "jest"
+    "build:lib": "babel ./src --out-dir ./lib && cpx './src/**/*.css' ./lib",
+    "build:umd": "cross-env NODE_ENV=development webpack --config ./webpack.config.js --colors --profile --progress src/index.js dist/react-sheets.js",
+    "build:umd:min": "cross-env NODE_ENV=production webpack --config ./webpack.config.js --colors --profile --progress src/index.js dist/react-sheets.min.js",
+    "build": "npm run build:lib && npm run build:umd && npm run build:umd:min",
+    "css": "cpx './src/**/*.css' ./lib",
+    "clean": "rimraf ./lib ./dist",
+    "rebuild": "npm run clean && npm run build",
+    "lint": "npm run lint:js && npm run lint:css",
+    "lint:css": "stylelint ./src/**/*.css",
+    "lint:js": "eslint ./src --ext .js",
+    "test": "cross-env NODE_ENV=test karma start",
+    "test:watch": "cross-env NODE_ENV=test karma start --single-run",
+    "example": "cross-env NODE_ENV=development webpack-dev-server --config ./example/webpack.config.js --content-base ./example --hot --inline --progress --display-error-details"
   },
-  "jest": {
-    "scriptPreprocessor": "<rootDir>/preprocessor.js",
-    "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/*"
-    ]
-  },
-  "browserify-shim": {
-    "react": "global:React",
-    "react/addons": "global:React",
-    "react-dom": "global:ReactDOM"
-  },
-  "browserify": {
-    "transform": [
-      "browserify-shim"
-    ]
-  },
-  "main": "./lib/spreadsheet.js",
-  "standalone": "React-Spreadsheet"
+  "files": [
+    "dist",
+    "lib",
+    "src"
+  ],
+  "main": "./lib/index.js"
 }
